help: fullscreen is gone
It was broken in 3.8.  Fixing is trivial, but then if you don't know the
keyboard shortcut, you'd be stuck in fullscreen (since there's now no
menubar or toolbar).  I think we'd need an overlay unfullscreen button,
like Totem, if we want to get it working again.  (But frankly, it
doesn't look very good fullscreen; I think we're better off without
it....)