Revert "Add a new game button to the scores dialog"
This reverts commit 6381c1692fa04add436ac42042377c7466232772.

Was added because I had removed the new game button from the window
locally. I deleted that commit locally but forgot to remove this one.

Don't put it back without teaching it not to appear when opening the
scores dialog from the app menu. (And probably just don't put it back.)